comparator investigation improved

+++ lab_07_3/sort.h
@@ -8,7 +8,7 @@ int int_comp(const void *int1, const void *int2);
 void el_copy(char *const what, char *const where, int size);
 int *binary_seek(char *left_el, char *right_el, const char *const comp_el, comparator, int size);
 void move_right(const int *const left_el, int *right_el, const int step);
-void mysort(void *start, int len, int size, comparator);
+void mysort(void *start, size_t len, size_t size, comparator);
 
 #endif
